在线更新修复
修复file.json和update.js因为没有在file.json中记录导致更新后文件被删除的bug
调整打包文件逻辑为不读取info.json和file.json，但每次更新会带上这两个文件

+++ js/config.js
@@ -142,6 +142,7 @@ export let config = {
 				this.textContent = '备份完毕，载入新内容中...';
 				await copyFiles('extension/活动武将/update_temp', 'extension/活动武将', needDownload);
 				for (const file of localFiles) {
+					if (new Set(['info.json', 'js/file.json']).has(file)) continue;
 					if (!remoteSet.has(file)) await game.promises.removeFile(`extension/活动武将/${file}`).catch(() => { });
 				}
 				//清理缓存
